refactor folder structure for `equality and diversity` section
