+++ integration_tests/pages/apply/about_the_person/equality_and_diversity/index.ts
@@ -1,0 +1,23 @@
+import AsianBackgroundPage from './asianBackgroundPage'
+import BlackBackgroundPage from './blackBackgroundPage'
+import DisabilityPage from './disabilityPage'
+import EthnicGroupPage from './ethnicGroupPage'
+import MixedBackgroundPage from './mixedBackgroundPage'
+import OtherBackgroundPage from './otherBackgroundPage'
+import SexAndGenderPage from './sexAndGenderPage'
+import SexualOrientationPage from './sexualOrientationPage'
+import WhiteBackgroundPage from './whiteBackgroundPage'
+import WillAnswerEqualityQuestionsPage from './willAnswerEqualityQuestionsPage'
+
+export {
+  AsianBackgroundPage,
+  BlackBackgroundPage,
+  DisabilityPage,
+  EthnicGroupPage,
+  MixedBackgroundPage,
+  OtherBackgroundPage,
+  SexAndGenderPage,
+  SexualOrientationPage,
+  WhiteBackgroundPage,
+  WillAnswerEqualityQuestionsPage,
+}
