+++ integration_tests/pages/apply/about_the_person/equality_and_diversity/sexualOrientationPage.ts
@@ -1,5 +1,5 @@
-import { Cas2Application as Application } from '../../../server/@types/shared/models/Cas2Application'
-import ApplyPage from './applyPage'
+import { Cas2Application as Application } from '../../../../../server/@types/shared/models/Cas2Application'
+import ApplyPage from '../../applyPage'
 
 export default class SexualOrientationPage extends ApplyPage {
   constructor(private readonly application: Application) {
