+++ integration_tests/tests/apply/about_the_person/equality_and_diversity/complete_other_background.cy.ts
@@ -11,7 +11,7 @@
 
 import Page from '../../../../pages/page'
 import TaskListPage from '../../../../pages/apply/taskListPage'
-import OtherBackgroundPage from '../../../../pages/apply/otherBackgroundPage'
+import OtherBackgroundPage from '../../../../pages/apply/about_the_person/equality_and_diversity/otherBackgroundPage'
 import { personFactory, applicationFactory } from '../../../../../server/testutils/factories/index'
 
 context('Visit "About the person" section', () => {
