+++ integration_tests/tests/apply/about_the_person/equality_and_diversity/complete_sex_gender_page.cy.ts
@@ -9,8 +9,10 @@
 //    Then I am taken to the 'Sexual orientation' page
 
 import Page from '../../../../pages/page'
-import SexualOrientationPage from '../../../../pages/apply/sexualOrientationPage'
-import SexAndGenderPage from '../../../../pages/apply/sexAndGenderPage'
+import {
+  SexAndGenderPage,
+  SexualOrientationPage,
+} from '../../../../pages/apply/about_the_person/equality_and_diversity'
 import { personFactory, applicationFactory } from '../../../../../server/testutils/factories/index'
 
 context('Visit "About the person" section', () => {
