+++ integration_tests/tests/apply/about_the_person/equality_and_diversity/complete_white_background_page.cy.ts
@@ -11,7 +11,7 @@
 
 import Page from '../../../../pages/page'
 import TaskListPage from '../../../../pages/apply/taskListPage'
-import WhiteBackgroundPage from '../../../../pages/apply/whiteBackgroundPage'
+import WhiteBackgroundPage from '../../../../pages/apply/about_the_person/equality_and_diversity/whiteBackgroundPage'
 import { personFactory, applicationFactory } from '../../../../../server/testutils/factories/index'
 
 context('Visit "About the person" section', () => {
